Refactor BytecodeExpressionVisitor to RowExprssionCompiler
