+++ presto-main/src/main/java/com/facebook/presto/sql/gen/BytecodeGeneratorContext.java
@@ -29,7 +29,7 @@ import static java.util.Objects.requireNonNull;
 
 public class BytecodeGeneratorContext
 {
-    private final BytecodeExpressionVisitor bytecodeGenerator;
+    private final RowExpressionCompiler rowExpressionCompiler;
     private final Scope scope;
     private final CallSiteBinder callSiteBinder;
     private final CachedInstanceBinder cachedInstanceBinder;
@@ -37,19 +37,19 @@ public class BytecodeGeneratorContext
     private final Variable wasNull;
 
     public BytecodeGeneratorContext(
-            BytecodeExpressionVisitor bytecodeGenerator,
+            RowExpressionCompiler rowExpressionCompiler,
             Scope scope,
             CallSiteBinder callSiteBinder,
             CachedInstanceBinder cachedInstanceBinder,
             FunctionRegistry registry)
     {
-        requireNonNull(bytecodeGenerator, "bytecodeGenerator is null");
+        requireNonNull(rowExpressionCompiler, "bytecodeGenerator is null");
         requireNonNull(cachedInstanceBinder, "cachedInstanceBinder is null");
         requireNonNull(scope, "scope is null");
         requireNonNull(callSiteBinder, "callSiteBinder is null");
         requireNonNull(registry, "registry is null");
 
-        this.bytecodeGenerator = bytecodeGenerator;
+        this.rowExpressionCompiler = rowExpressionCompiler;
         this.scope = scope;
         this.callSiteBinder = callSiteBinder;
         this.cachedInstanceBinder = cachedInstanceBinder;
@@ -69,7 +69,7 @@ public class BytecodeGeneratorContext
 
     public BytecodeNode generate(RowExpression expression)
     {
-        return expression.accept(bytecodeGenerator, scope);
+        return rowExpressionCompiler.compile(expression, scope);
     }
 
     public FunctionRegistry getRegistry()
